Add files via upload
In this project, we developed an efficient and simple method to predict antimicrobial peptides (AMPs) 
activity from their amino acid sequences. This prediction aids in drug design by streamlining the lab 
work of pharmacists. Utilizing an extensive dataset from different resources, one of which was a 
collaborative work with the results of a research study by Dr. Ahmad Al Omari on big data bot, which 
included 1,295 sequences of both AMPs and non-AMPs. we investigated different peptide features such 
as amino acid composition, physicochemical properties and amino acid index and many more then tried 
to train the model to find correlations between the features frequencies and peptide activity to help in the 
prediction process, using the Random Forest model, we achieved an accuracy of 92% and after many 
trials we decided to apply Principal Component Analysis (PCA) and normalization with the data 
preprocessing step then we were able to enhance the accuracy to 95%. Additionally, we made use of 
NVIDIA's new technology CUDF to accelerate pandas performance. We achieved the end goal of a 
robust machine learning model that significantly reduces the complexity and time required for AMP 
classification, providing a valuable tool for pharmaceutical research and development in drug design that 
can accelerate the discovery and development of new, effective antibiotics.